• much nicer table output :)
I only changed the MySQL query command, but the PostgreSQL should naturally follow. Either someone send me the HTML output from PostgreSQL (so that I can see how to alter it, to add a class to the table), or someone with PostgreSQL add the proper replacement (and inclusion of our “default” style sheet).



git-svn-id: http://svn.textmate.org/trunk/Support@2731 dfb7d73b-c2ec-0310-8fea-fb051d288c6d