• the selected style (bright, dark, no colors) is now remembered
Unfortunately with the dark theme the background is white for a fraction of a second. It might be possible to improve this by trimming some of the CSS inserted before WebKit reaches the body tag with style info, but likely I will need to add some special support for this in TextMate, like delay showing the window before the body tag has been seen or 0.1 second has passed (whatever comes first) — my fear is that this may result in an overall degradation of the output window, e.g. for commands which do not set the body (which will then suffer the full delay) and also just not seeming that responsive… OTOH if you use a nocturnal theme on a 23" display a white flash is a little hard on the eyes.


git-svn-id: http://svn.textmate.org/trunk/Support@4912 dfb7d73b-c2ec-0310-8fea-fb051d288c6d