• Update my_popen3 to utilise tm_interactive_input.dylib if the library is present.
This brings interactive input capabilities to all processes using scriptmate (and my_popen3) as it's basis.

TM_INTERACTIVE_INPUT is set to 'AUTO|ECHO' which manifests as the requester being invoked when a request is made for input and there is none already waiting, and entered input is echoed to output when it is captured.

git-svn-id: http://svn.textmate.org/trunk/Support@9317 dfb7d73b-c2ec-0310-8fea-fb051d288c6d