• New file that provides TextMate::Process.run() which can be used to open processes and capture their output.
The method takes a nested array of shell words to form the command used to launch the process. The output can either be collected as the process executes by providing a block, or collected as the return of the method.

The last non block argument can also be a hash of options that will augment the behavior. See the doc header of the file for details.

git-svn-id: http://svn.textmate.org/trunk/Support@9630 dfb7d73b-c2ec-0310-8fea-fb051d288c6d