• Trap INT and TERM (as well as USR1) so we give the process we launched the opportunity to shutdown cleanly.
We were previously doing this but it was deactivated by Alex Ross. It seems that Alex assumed that it wasn't required as Executor provides a pseudo ⌃C feature which sends USR1 to the parent process which we interpret as a request to kill the child that we launched. However, this doesn't extend to processes NOT launched through Executor so we DO need to trap and forward INT and TERM.

git-svn-id: http://svn.textmate.org/trunk/Support@11600 dfb7d73b-c2ec-0310-8fea-fb051d288c6d