Use correct encoding for man pages
We generate UTF-8 man pages and set ‘charset=UTF-8’ using the <meta> tag, but WebKit ignores this.

I also tried to set the ‘com.apple.TextEncoding’ extended attribute (to UTF-8), but this is ignored as well.

So the only fix was to include a “byte order mark” (BOM) in the file.