TextMate.executor now parses «file»:«line» references
It will turn these into links to the file and properly handle the case where «file» is pointing to a temporary file (because the document is untitled).

Additionally, if the current file is unsaved or read-only (but has been saved to a temporary file), it will convert txmt://-references to this temporary file into references to the current document and likewise convert references to the temporary file’s name to the current document’s display name.