Better hyperlinking of untitled documents
Some interpreters will resolve symbolic links in the path they are given, and use this resolved (“real”) path in the error output (e.g. php and coffee-script).

When running untitled documents, we previously only searched for the non-resolved path in the error output (for hyperlinking).

This issue only applies to code using TextMate.executor that does its own hyperlinking.