Using ⌃C to break command execution could throw exception
This would happen if the direct command executed terminates but a child keeps stdout/err open, causing TextMate to wait for the child to finish.

Pressing ⌃C will search for the process ID of the initial process. If this no longer exist, nil is found and subsequently used, causing a “no method” exception.

The reported case was using the Make bundle with a rule that calls ‘@@mongod --fork’.