Use Ruby 1.8 for compatibility on Yosemite
This is for backwards compatibility with bundle items that were designed for ruby 1.8. Input is passed along to ruby 1.8 if already installed or if needed downloading and installing a copy for TextMates exclusive use. This does not affect the system ruby version or any other part of the system. Revised from previous commit due to an error in the ruby build.