Executor output: Highlight line if file reference link is clicked.
Motivation: Make it easier to keep track of the file reference link you just clicked on: You click on a link, focus changes to the associated source document, and now you’re wondering: “What’s the exact error message associated with this line again?” If the output window contains a lot of links it can be hard to find the exact one you just clicked to read the associated message.

I experimented with highlighting the link only, but highlighting the whole line somehow works better (especially for quickly looking at the associated message).