Replace Makefile with rave definition for building find_app
With universal builds, code signing, and notarization, it is not really practical to manually maintain a Makefile to do all these things.

I may add a Makefile generator to ‘rave’ so that we can include a (generated) Makefile. The reason I do not include the generated ‘build.ninja’ is because signing and notarization values are tied to developer ID, and ninja does not support reading values from environment variables.